Added GameObject Class + documentation
The core of the gameobject class was added, along with a basic Vector calss to support it. I also Isolated the engine from the game by placing it in a folder. Unfortunately this meant I had to add more js files. to index.html. 

Some of the gamecode is still in the engine files, and I will isolate that by introducing user made levels. 

I also added documentation in the form  of Typedoc.