Added window resizing and shifted to outfile
The game window now auto-resizes to the browser, but it's alwys slightly too big. I think it's the padding on the html. I'll have to either compensate for it or remove it.

I switched the module system to AMD because it supports cating all the output js into a single file. It seems to be working with atom-typescript, but tsc is outputting errors in the build.  Atom's also having an issue where it thinks th js is always out of date.