Added Gameloop, levels, and supporting constructs
The game object now runs a gameloop, which just hooks the logic and rendering of the current level.  

Levels were always intended to be the developer's entry point to the game. THey are greated by the developer and passed to the game object.  They are a subclass of groups, which are just collections of gameobjects.

The concept of the gameobject has been generalized to the entity, of which both gameobject and group are a subclass. This allows the group clas to avoid meaningless overhead, like velocity. I probably should have commited before doing that refoactor, but I forgot.